Organize example scripts into hierarchical packages
